+++ server.js
@@ -1,6 +1,6 @@
 /**
  * AIMapa - Server
- * Verze 0.3.5.2
+ * Verze 0.3.5.3
  */
 
 // Načtení modulů